requirements.txt
numpy==1.23.5
scipy==1.10.0
pandas==1.5.2
mat73==0.63
pynrrd==1.0.0
scikit-image==0.19.3
scikit-learn==1.2.2
seaborn==0.12.2
Pillow==9.3.0
nibabel==5.1.0
openpyxl==3.1.4
